+++ scripts/tests/-0Executed as native process.ps1
@@ -1,4 +1,4 @@
-#v1.03
+#v1.04
 #https://github.com/texhex/testutil/wiki/_fwLinkScript
 
 
@@ -9,4 +9,4 @@
 Set-StrictMode -version 2.0
 
 #Terminate script on errors 
-$ErrorActionPreference = 'Stop'#Default is OK because we expect that Test-TUWoW returns false$Return = @{Data = "OK"; Name="Executed with correct bitness"; Text= "TestUtil is executed with the correct bitness (as 64 bit on a 64 bit OS or 32 bit on a 32 bit OS)"}if (Test-TUWoW) {       #We are affected by WoW64. Report this as Fail (MAJOR issue)   $Return.Data="Fail"   $Return.Text="TestUtil was not started with correct bitness! Please start it using TestUtilLauncher.exe."}$Return
+$ErrorActionPreference = 'Stop'#Default is OK because we expect that Test-TUWoW returns false$Result = @{Data = "OK"; Name="Executed with correct bitness"; Text= "TestUtil is executed with the correct bitness (as 64 bit on a 64 bit OS or 32 bit on a 32 bit OS)"}if (Test-MPXWoW) {       #We are affected by WoW64. Report this as Fail (MAJOR issue)   $Result.Data="Fail"   $Result.Text="TestUtil was not started with correct bitness! Please start it using TestUtilLauncher.exe."}$Result