+++ scripts/tests/-Free Memory available.ps1
@@ -1,4 +1,4 @@
-#v1.02
+#v1.03
 #https://github.com/texhex/testutil/wiki/_fwLinkScript
 
 
@@ -9,4 +9,4 @@
 Set-StrictMode -version 2.0
 
 #Terminate script on errors 
-$ErrorActionPreference = 'Stop'$Return = @{Data = "OK"; Name="Minimum free physical memory"; Text= "More than 256 MB of physical memory is available"}$wmi = Get-WMIObject Win32_OperatingSystem -Property "FreePhysicalMemory","FreeVirtualMemory"[int]$memFreePhyscial = $wmi.FreePhysicalMemory #in KB[int]$memFreeVirtual = $wmi.FreeVirtualMemory #in KB$memFreePhyscial=$memFreePhyscial/1024 #now in MB$memFreeVirtual=$memFreeVirtual/1024 #now in MB#Debug#$memFreePhyscial=10#$memFreeVirtual#If less than 256MB are available, report Majorif ($memFreePhyscial -lt 256) {   $Return.Data="Fail"   $Return.Text="There are only $memFreePhyscial MB of free physical memory available"}$Return
+$ErrorActionPreference = 'Stop'$Result = @{Data = "OK"; Name="More than 256 MB of free physical memory";}$wmi = Get-WMIObject Win32_OperatingSystem -Property "FreePhysicalMemory","FreeVirtualMemory"[int]$memFreePhyscial = $wmi.FreePhysicalMemory #in KB[int]$memFreeVirtual = $wmi.FreeVirtualMemory #in KB"Free memory in KB: $memFreePhyscial"$memFreePhyscial_human=Get-MPXHumanizeBytes ($memFreePhyscial*1024) #because the value is in KB#append the freespace to the text of this test$Result.Text="$memFreePhyscial_human of free phyisical memory available"$memFreePhyscial=$memFreePhyscial/1024 #now in MB$memFreeVirtual=$memFreeVirtual/1024 #now in MB"Free memory in MB: $memFreePhyscial"#Debug#$memFreePhyscial=10#$memFreeVirtual#If less than 256MB are available, report Majorif ($memFreePhyscial -lt 256) {   $Result.Data="Fail"}$Result